Madgwick filter functioning, but not useful during flight
