+++ flight_src/mcu/lib/teensyShared/MadgwickAHRS.cpp
@@ -1,5 +1,5 @@
 /**
- * @file        MadgwickAHRS.h
+ * @file        MadgwickAHRS.cpp
  * @authors     Ayberk Yaraneri
  *
  * @brief   A wrapper class that implements the Madwick AHRS filter for attitude estimation.
@@ -11,54 +11,253 @@
  *
  */
 
-#ifndef MADGWICK_AHRS_H
-#define MADGWICK_AHRS_H
+#include "MadgwickAHRS.h"
 
-#include "sensors.h"
+#include <iostream>
 
-#include "GlobalVars.h"
+#include <math.h>
 
-#define sampleFreq	100.0f		// sample frequency in Hz
-#define betaDef		0.1f		// 2 * proportional gain
+void MadgwickAHRS::MadgwickAHRSupdate() {
 
-class MadgwickAHRS {
-    public:
-        MadgwickAHRS(pointers* ptr) {
-            pointer_struct_ = ptr;
+    // Fetch data from sensors
+    chMtxLock(dataMutex_lowG_);
+    float gx = lowG_data_ptr_->gx;
+    float gy = lowG_data_ptr_->gy;
+    float gz = lowG_data_ptr_->gz;
 
-            state_data_ptr_ = &pointer_struct_->stateData;
-            lowG_data_ptr_ = &pointer_struct_->sensorDataPointer->lowG_data;
-            dataMutex_lowG_ = &pointer_struct_->dataloggerTHDVarsPointer.dataMutex_lowG;
-            dataMutex_state_ = &pointer_struct_->dataloggerTHDVarsPointer.dataMutex_state;
+    float ax = lowG_data_ptr_->ax;
+    float ay = lowG_data_ptr_->ay;
+    float az = lowG_data_ptr_->az;
 
-            // SILSIM Data Logging
-            madgwick_logger_ = std::make_shared<spdlog::logger>("MadgwickAHRS", silsim_datalog_sink);
-            madgwick_logger_->info("DATALOG_FORMAT," + datalog_format_string);
-        };
+    float mx = lowG_data_ptr_->mx;
+    float my = lowG_data_ptr_->my;
+    float mz = lowG_data_ptr_->mz;
+    chMtxUnlock(dataMutex_lowG_);
 
-        void MadgwickAHRSupdate();
-        void MadgwickAHRSupdateIMU();
 
-        // SILSIM Data Logging
-        void log_madgwick_state(double tStamp);
+	float recipNorm;
+	float s0, s1, s2, s3;
+	float qDot1, qDot2, qDot3, qDot4;
+	float hx, hy;
+	float _2q0mx, _2q0my, _2q0mz, _2q1mx, _2bx, _2bz, _4bx, _4bz, _2q0, _2q1, _2q2, _2q3, _2q0q2, _2q2q3, q0q0, q0q1, q0q2, q0q3, q1q1, q1q2, q1q3, q2q2, q2q3, q3q3;
 
-    private:
-        pointers* pointer_struct_;
-        StateData* state_data_ptr_;
-        LowGData* lowG_data_ptr_;
+	// Use IMU algorithm if magnetometer measurement invalid (avoids NaN in magnetometer normalisation)
+	if((mx == 0.0f) && (my == 0.0f) && (mz == 0.0f)) {
+		MadgwickAHRSupdateIMU();
+		return;
+	}
 
-        mutex_t* dataMutex_lowG_;
-        mutex_t* dataMutex_state_;
+	// Rate of change of quaternion from gyroscope
+	qDot1 = 0.5f * (-q1 * gx - q2 * gy - q3 * gz);
+	qDot2 = 0.5f * (q0 * gx + q2 * gz - q3 * gy);
+	qDot3 = 0.5f * (q0 * gy - q1 * gz + q3 * gx);
+	qDot4 = 0.5f * (q0 * gz + q1 * gy - q2 * gx);
 
-        float beta_ = betaDef;
-        float q0 = 1.0, q1 = 0.0, q2 = 0.0, q3 = 0.0;
+	// Compute feedback only if accelerometer measurement valid (avoids NaN in accelerometer normalisation)
+	if(!((ax == 0.0f) && (ay == 0.0f) && (az == 0.0f))) {
 
-        float invSqrt(float x);
+		// Normalise accelerometer measurement
+		recipNorm = invSqrt(ax * ax + ay * ay + az * az);
+		ax *= recipNorm;
+		ay *= recipNorm;
+		az *= recipNorm;   
 
-        // SILSIM Data Logging
-        std::shared_ptr<spdlog::logger> madgwick_logger_;
-        std::string datalog_format_string = 
-            "timestamp,q0,q1,q2,q3";
-};
+		// Normalise magnetometer measurement
+		recipNorm = invSqrt(mx * mx + my * my + mz * mz);
+		mx *= recipNorm;
+		my *= recipNorm;
+		mz *= recipNorm;
 
-#endif // MADGWICK_AHRS_H
+		// Auxiliary variables to avoid repeated arithmetic
+		_2q0mx = 2.0f * q0 * mx;
+		_2q0my = 2.0f * q0 * my;
+		_2q0mz = 2.0f * q0 * mz;
+		_2q1mx = 2.0f * q1 * mx;
+		_2q0 = 2.0f * q0;
+		_2q1 = 2.0f * q1;
+		_2q2 = 2.0f * q2;
+		_2q3 = 2.0f * q3;
+		_2q0q2 = 2.0f * q0 * q2;
+		_2q2q3 = 2.0f * q2 * q3;
+		q0q0 = q0 * q0;
+		q0q1 = q0 * q1;
+		q0q2 = q0 * q2;
+		q0q3 = q0 * q3;
+		q1q1 = q1 * q1;
+		q1q2 = q1 * q2;
+		q1q3 = q1 * q3;
+		q2q2 = q2 * q2;
+		q2q3 = q2 * q3;
+		q3q3 = q3 * q3;
+
+		// Reference direction of Earth's magnetic field
+		hx = mx * q0q0 - _2q0my * q3 + _2q0mz * q2 + mx * q1q1 + _2q1 * my * q2 + _2q1 * mz * q3 - mx * q2q2 - mx * q3q3;
+		hy = _2q0mx * q3 + my * q0q0 - _2q0mz * q1 + _2q1mx * q2 - my * q1q1 + my * q2q2 + _2q2 * mz * q3 - my * q3q3;
+		_2bx = sqrt(hx * hx + hy * hy);
+		_2bz = -_2q0mx * q2 + _2q0my * q1 + mz * q0q0 + _2q1mx * q3 - mz * q1q1 + _2q2 * my * q3 - mz * q2q2 + mz * q3q3;
+		_4bx = 2.0f * _2bx;
+		_4bz = 2.0f * _2bz;
+
+		// Gradient decent algorithm corrective step
+		s0 = -_2q2 * (2.0f * q1q3 - _2q0q2 - ax) + _2q1 * (2.0f * q0q1 + _2q2q3 - ay) - _2bz * q2 * (_2bx * (0.5f - q2q2 - q3q3) + _2bz * (q1q3 - q0q2) - mx) + (-_2bx * q3 + _2bz * q1) * (_2bx * (q1q2 - q0q3) + _2bz * (q0q1 + q2q3) - my) + _2bx * q2 * (_2bx * (q0q2 + q1q3) + _2bz * (0.5f - q1q1 - q2q2) - mz);
+		s1 = _2q3 * (2.0f * q1q3 - _2q0q2 - ax) + _2q0 * (2.0f * q0q1 + _2q2q3 - ay) - 4.0f * q1 * (1 - 2.0f * q1q1 - 2.0f * q2q2 - az) + _2bz * q3 * (_2bx * (0.5f - q2q2 - q3q3) + _2bz * (q1q3 - q0q2) - mx) + (_2bx * q2 + _2bz * q0) * (_2bx * (q1q2 - q0q3) + _2bz * (q0q1 + q2q3) - my) + (_2bx * q3 - _4bz * q1) * (_2bx * (q0q2 + q1q3) + _2bz * (0.5f - q1q1 - q2q2) - mz);
+		s2 = -_2q0 * (2.0f * q1q3 - _2q0q2 - ax) + _2q3 * (2.0f * q0q1 + _2q2q3 - ay) - 4.0f * q2 * (1 - 2.0f * q1q1 - 2.0f * q2q2 - az) + (-_4bx * q2 - _2bz * q0) * (_2bx * (0.5f - q2q2 - q3q3) + _2bz * (q1q3 - q0q2) - mx) + (_2bx * q1 + _2bz * q3) * (_2bx * (q1q2 - q0q3) + _2bz * (q0q1 + q2q3) - my) + (_2bx * q0 - _4bz * q2) * (_2bx * (q0q2 + q1q3) + _2bz * (0.5f - q1q1 - q2q2) - mz);
+		s3 = _2q1 * (2.0f * q1q3 - _2q0q2 - ax) + _2q2 * (2.0f * q0q1 + _2q2q3 - ay) + (-_4bx * q3 + _2bz * q1) * (_2bx * (0.5f - q2q2 - q3q3) + _2bz * (q1q3 - q0q2) - mx) + (-_2bx * q0 + _2bz * q2) * (_2bx * (q1q2 - q0q3) + _2bz * (q0q1 + q2q3) - my) + _2bx * q1 * (_2bx * (q0q2 + q1q3) + _2bz * (0.5f - q1q1 - q2q2) - mz);
+		recipNorm = invSqrt(s0 * s0 + s1 * s1 + s2 * s2 + s3 * s3); // normalise step magnitude
+		s0 *= recipNorm;
+		s1 *= recipNorm;
+		s2 *= recipNorm;
+		s3 *= recipNorm;
+
+		// Apply feedback step
+		qDot1 -= beta_ * s0;
+		qDot2 -= beta_ * s1;
+		qDot3 -= beta_ * s2;
+		qDot4 -= beta_ * s3;
+	}
+
+	// Integrate rate of change of quaternion to yield quaternion
+	q0 += qDot1 * (1.0f / sampleFreq);
+	q1 += qDot2 * (1.0f / sampleFreq);
+	q2 += qDot3 * (1.0f / sampleFreq);
+	q3 += qDot4 * (1.0f / sampleFreq);
+
+	// Normalise quaternion
+	recipNorm = invSqrt(q0 * q0 + q1 * q1 + q2 * q2 + q3 * q3);
+	q0 *= recipNorm;
+	q1 *= recipNorm;
+	q2 *= recipNorm;
+	q3 *= recipNorm;
+
+
+    // Write orientation estimate to stateData struct
+    chMtxLock(dataMutex_state_);
+    state_data_ptr_->state_q0 = q0;
+    state_data_ptr_->state_q1 = q1;
+    state_data_ptr_->state_q2 = q2;
+    state_data_ptr_->state_q3 = q3;
+    state_data_ptr_->timeStamp_state = chVTGetSystemTime();
+    chMtxUnlock(dataMutex_state_);
+}
+
+
+void MadgwickAHRS::MadgwickAHRSupdateIMU() {
+
+    // Fetch data from sensors
+    chMtxLock(dataMutex_lowG_);
+    float gx = lowG_data_ptr_->gx;
+    float gy = lowG_data_ptr_->gy;
+    float gz = lowG_data_ptr_->gz;
+
+    float ax = lowG_data_ptr_->ax;
+    float ay = lowG_data_ptr_->ay;
+    float az = lowG_data_ptr_->az;
+    chMtxUnlock(dataMutex_lowG_);
+
+	float recipNorm;
+	float s0, s1, s2, s3;
+	float qDot1, qDot2, qDot3, qDot4;
+	float _2q0, _2q1, _2q2, _2q3, _4q0, _4q1, _4q2 ,_8q1, _8q2, q0q0, q1q1, q2q2, q3q3;
+
+	// Rate of change of quaternion from gyroscope
+	qDot1 = 0.5f * (-q1 * gx - q2 * gy - q3 * gz);
+	qDot2 = 0.5f * (q0 * gx + q2 * gz - q3 * gy);
+	qDot3 = 0.5f * (q0 * gy - q1 * gz + q3 * gx);
+	qDot4 = 0.5f * (q0 * gz + q1 * gy - q2 * gx);
+
+	// Compute feedback only if accelerometer measurement valid (avoids NaN in accelerometer normalisation)
+	if(!((ax == 0.0f) && (ay == 0.0f) && (az == 0.0f))) {
+
+		// Normalise accelerometer measurement
+		recipNorm = invSqrt(ax * ax + ay * ay + az * az);
+		ax *= recipNorm;
+		ay *= recipNorm;
+		az *= recipNorm;   
+
+		// Auxiliary variables to avoid repeated arithmetic
+		_2q0 = 2.0f * q0;
+		_2q1 = 2.0f * q1;
+		_2q2 = 2.0f * q2;
+		_2q3 = 2.0f * q3;
+		_4q0 = 4.0f * q0;
+		_4q1 = 4.0f * q1;
+		_4q2 = 4.0f * q2;
+		_8q1 = 8.0f * q1;
+		_8q2 = 8.0f * q2;
+		q0q0 = q0 * q0;
+		q1q1 = q1 * q1;
+		q2q2 = q2 * q2;
+		q3q3 = q3 * q3;
+
+		// Gradient decent algorithm corrective step
+		s0 = _4q0 * q2q2 + _2q2 * ax + _4q0 * q1q1 - _2q1 * ay;
+		s1 = _4q1 * q3q3 - _2q3 * ax + 4.0f * q0q0 * q1 - _2q0 * ay - _4q1 + _8q1 * q1q1 + _8q1 * q2q2 + _4q1 * az;
+		s2 = 4.0f * q0q0 * q2 + _2q0 * ax + _4q2 * q3q3 - _2q3 * ay - _4q2 + _8q2 * q1q1 + _8q2 * q2q2 + _4q2 * az;
+		s3 = 4.0f * q1q1 * q3 - _2q1 * ax + 4.0f * q2q2 * q3 - _2q2 * ay;
+		recipNorm = invSqrt(s0 * s0 + s1 * s1 + s2 * s2 + s3 * s3); // normalise step magnitude
+		s0 *= recipNorm;
+		s1 *= recipNorm;
+		s2 *= recipNorm;
+		s3 *= recipNorm;
+
+		// Apply feedback step
+		qDot1 -= beta_ * s0;
+		qDot2 -= beta_ * s1;
+		qDot3 -= beta_ * s2;
+		qDot4 -= beta_ * s3;
+	}
+
+	// Integrate rate of change of quaternion to yield quaternion
+	q0 += qDot1 * (1.0f / sampleFreq);
+	q1 += qDot2 * (1.0f / sampleFreq);
+	q2 += qDot3 * (1.0f / sampleFreq);
+	q3 += qDot4 * (1.0f / sampleFreq);
+
+	// Normalise quaternion
+	recipNorm = invSqrt(q0 * q0 + q1 * q1 + q2 * q2 + q3 * q3);
+	q0 *= recipNorm;
+	q1 *= recipNorm;
+	q2 *= recipNorm;
+	q3 *= recipNorm;
+
+
+    // Write orientation estimate to stateData struct
+    chMtxLock(dataMutex_state_);
+    state_data_ptr_->state_q0 = q0;
+    state_data_ptr_->state_q1 = q1;
+    state_data_ptr_->state_q2 = q2;
+    state_data_ptr_->state_q3 = q3;
+    state_data_ptr_->timeStamp_state = chVTGetSystemTime();
+    chMtxUnlock(dataMutex_state_);
+}
+
+
+// Fast inverse square-root
+// See: http://en.wikipedia.org/wiki/Fast_inverse_square_root
+
+float MadgwickAHRS::invSqrt(float x) {
+	float halfx = 0.5f * x;
+	float y = x;
+	long i = *(long*)&y;
+	i = 0x5f3759df - (i>>1);
+	y = *(float*)&i;
+	y = y * (1.5f - (halfx * y * y));
+	return y;
+}
+
+// SILSIM Data Logging
+void MadgwickAHRS::log_madgwick_state(double tStamp) {
+    if (madgwick_logger_) {
+
+        std::stringstream datalog_ss;
+
+        datalog_ss << "DATA,"
+                   << tStamp << ","
+                   << q0 << ","
+                   << q1 << ","
+                   << q2 << ","
+                   << q3; 
+
+        madgwick_logger_->info(datalog_ss.str());
+    }
+}
