+++ flight_src/mcu/lib/teensyShared/sensors.cpp
@@ -69,10 +69,6 @@ void lowGimuTickFunction(LSM9DS1* lsm, DataLogBuffer* data_log_buffer,
     lowG_Data->mz = lsm->calcMag(lsm->mz);
     //! Unlocking &dataMutex for low g
     
-    std::cout << "Debug lowGimuTickFunction: "; 
-    // std::cout << lsm->calcMag(lsm->mx) << "," << lsm->calcMag(lsm->my) << "," << lsm->calcMag(lsm->mz) << std::endl;
-    std::cout << lsm->mx << "," << lsm->my << "," << lsm->mz << std::endl;
-
     data_log_buffer->lowGFifo.push(*lowG_Data);
     chMtxUnlock(&data_log_buffer->dataMutex_lowG);
